Add option to specify default included resources (#250)
* Add option to specify default included resources

Let's provide an extra JSONAPIMeta configuration option that specifies
resources to be listed in the 'included' section of the response.
These resources shall be included even if they're not explicitly mentioned
in the 'include' request parameter.

* Re-trigger Travis Build script

The previous build failure was caused by some
random environment issue.

* Ignore default 'included' resources if specified in request

Conform to the specs and ensure that no other resources are included
in the response other than those explicitely requested inside the
'include' parameter in request.

* Add tests for default included_resources feature